test(portable-text-editor): move schema test file to test root
